requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.25
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
authlib==1.2.1
itsdangerous==2.1.2
httpx==0.25.2
tortoise3
bcrypt
passlib
pyjwt
pydantic
pydantic-settings
